[WIP] move state management into a not quite functioning custom hook
