+++ apps/src/templates/codeReviewGroups/CodeReviewGroupsManager.jsx
@@ -1,7 +1,7 @@
-import React, {useState} from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import {DragDropContext} from 'react-beautiful-dnd';
-import _ from 'lodash';
+import {useGroups} from './codeReviewGroupManagerHooks';
 import UnassignedStudentsPanel from './UnassignedStudentsPanel';
 import AssignedStudentsPanel from './AssignedStudentsPanel';
 
@@ -14,57 +14,17 @@ const DROPPABLE_ID_UNASSIGNED = 'unassigned';
 // can be found here:
 // https://github.com/atlassian/react-beautiful-dnd
 export default function CodeReviewGroupsManager({initialGroups}) {
-  const [groups, setGroups] = useState(
+  const [groups, getGroup, setGroups] = useGroups(
     initialGroups.map(group => addDroppableIdToGroup(group))
   );
 
-  const getGroup = droppableId =>
-    _.find(groups, group => group.droppableId === droppableId);
   const getUnassignedGroup = () => getGroup(DROPPABLE_ID_UNASSIGNED);
   const getAssignedGroups = () =>
     groups.filter(group => group.droppableId !== DROPPABLE_ID_UNASSIGNED);
 
-  function onDragEnd(result) {
-    const {source, destination} = result;
-    const sourceId = source.droppableId;
-
-    // dropped outside the group
-    if (!destination) {
-      return;
-    }
-
-    const destinationId = destination.droppableId;
-
-    if (sourceId === destinationId) {
-      // If a member was dropped in its current group.
-      const result = reorder(
-        getGroup(sourceId),
-        source.index,
-        destination.index
-      );
-
-      const updatedGroups = updateGroups(groups, [result]);
-      setGroups(updatedGroups);
-    } else {
-      // If a member was dropped in a different group.
-      const result = move(
-        getGroup(sourceId),
-        getGroup(destinationId),
-        source.index,
-        destination.index
-      );
-
-      const updatedGroups = updateGroups(groups, [
-        result.updatedSource,
-        result.updatedDest
-      ]);
-
-      setGroups(updatedGroups);
-    }
-  }
-
+  // Creating groups no longer works (b/c we want to set groups here...)
   return (
-    <DragDropContext onDragEnd={onDragEnd}>
+    <DragDropContext onDragEnd={setGroups}>
       <div style={styles.dragAndDropContainer}>
         <UnassignedStudentsPanel unassignedGroup={getUnassignedGroup()} />
         <AssignedStudentsPanel
@@ -80,48 +40,6 @@ export default function CodeReviewGroupsManager({initialGroups}) {
 
 CodeReviewGroupsManager.propTypes = {
   initialGroups: PropTypes.array.isRequired
-};
-
-// Reorders members in a group if member dragged elsewhere in the same group.
-// Returns a copied, updated group.
-const reorder = (group, startIndex, endIndex) => {
-  const result = {...group};
-  const [removed] = result.members.splice(startIndex, 1);
-  result.members.splice(endIndex, 0, removed);
-
-  return result;
-};
-
-// Moves an member from one group to another group.
-// Returns copies of both updated groups.
-const move = (
-  source,
-  destination,
-  droppableSourceIndex,
-  droppableDestinationIndex
-) => {
-  const updatedSource = {...source};
-  const updatedDest = {...destination};
-  const [removed] = updatedSource.members.splice(droppableSourceIndex, 1);
-
-  updatedDest.members.splice(droppableDestinationIndex, 0, removed);
-
-  return {updatedSource, updatedDest};
-};
-
-// Returns a copied list of groups
-// with updated versions of the provided changedGroups (ie, with more, less, or reordered members).
-const updateGroups = (groups, changedGroups) => {
-  const updatedGroups = [...groups];
-
-  changedGroups.forEach(changedGroup => {
-    const updatedGroupIndex = updatedGroups.findIndex(
-      group => group.droppableId === changedGroup.droppableId
-    );
-    updatedGroups[updatedGroupIndex] = changedGroup;
-  });
-
-  return updatedGroups;
 };
 
 const addDroppableIdToGroup = group => {
